*** empty log message ***
git-svn-id: http://svn.neuromancy.net/magick/trunk/Magick-II@342 1d68a2a4-f605-0410-a2a0-863aa388aa54

+++ include/server.h
@@ -24,8 +24,8 @@ class Server
     mstring i_Uplink;
     int i_Hops;
     mstring i_Description;
-    double i_Ping;
-    double i_Lag;
+    unsigned long i_Ping;
+    unsigned long i_Lag;
 public:
     Server();
     Server(const Server &in);
